Adding Saros_Session and Saros_Auth to the party. Session should be operational, Auth needs more work with authenticating against Spot. I'm currently unsure of how to handle the following situation:
A users table:
id, username, password

auth validates a username, password pair. If it is successfull, it stores the username in the session. ACL then uses the string of the username for permissions. Ideally ACL would use the ID.

Possible Solution: Store a resulting object in the session, instead of just a string